+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/TextComponents.java
@@ -21,18 +21,12 @@ package is.codion.swing.common.ui.component.text;
 import javax.swing.JTextField;
 import javax.swing.text.AbstractDocument;
 import javax.swing.text.Document;
-import javax.swing.text.JTextComponent;
 import java.awt.Dimension;
-import java.awt.event.FocusListener;
-
-import static java.util.Objects.requireNonNull;
 
 /**
  * A utility class for JTextComponents.
  */
 public final class TextComponents {
-
-	private static final String TEXT_COMPONENT = "textComponent";
 
 	private static Dimension preferredTextFieldSize;
 
@@ -63,38 +57,6 @@ public final class TextComponents {
 	 */
 	public static void lowerCase(Document document) {
 		new TextFieldDocumentCase(document, CaseDocumentFilter.DocumentCase.LOWERCASE);
-	}
-
-	/**
-	 * Selects all text in the given component when it gains focus and clears
-	 * the selection when focus is lost
-	 * @param textComponent the text component
-	 * @param <T> the component type
-	 * @return the component
-	 */
-	public static <T extends JTextComponent> T selectAllOnFocusGained(T textComponent) {
-		requireNonNull(textComponent, TEXT_COMPONENT);
-		textComponent.addFocusListener(new SelectAllFocusListener(textComponent));
-
-		return textComponent;
-	}
-
-	/**
-	 * Reverts the functionality added via {@link #selectAllOnFocusGained(JTextComponent)}.
-	 * @param textComponent the text component
-	 * @param <T> the component type
-	 * @return the text component
-	 * @see #selectAllOnFocusGained(JTextComponent)
-	 */
-	public static <T extends JTextComponent> T selectNoneOnFocusGained(T textComponent) {
-		requireNonNull(textComponent, TEXT_COMPONENT);
-		for (FocusListener listener : textComponent.getFocusListeners()) {
-			if (listener instanceof SelectAllFocusListener) {
-				textComponent.removeFocusListener(listener);
-			}
-		}
-
-		return textComponent;
 	}
 
 	/**